Better way to handle logging
Instead of configuring the global root logger and attaching a default
handler to it (which library users may not want) we give each module
its own logger using the standard `logging.getLogger(__name__)`
convention.

This allows library users to configure as much or as little logging as
they want and allows them to attach whatever handlers they want.

Signed-off-by: Alexander Hermes <alexander.hermes@grasshopperasia.com>